Update dependencies for the project
I started to use `coverage` to track the lines of the uncovered code
with my scripts. This happened on the day 10, when was working on the
second solution. Yes, I'm aware that code coverage says nothing about
code's quality, but I use it to discover the lines of code that require
my attention.

Also, I decided to add linters and formater to the list of requirements, to
track what tools I'm using while working on solutions.